ISSUE #2695 - Handle links passed by props
